requirements.txt
Django~=3.2.9
Pillow~=8.4.0
factory_boy~=3.2.1